rain data now in utilities dir; called in server but currently commented out
