+++ server/server.js
@@ -7,11 +7,11 @@ const trails = require('./routes/trails');
 const Trail = require('./db/models/Trails');
 const rp = require('request-promise');
 const weatherKey = require('../config/config');
-const rain = require('./routes/rain');
 const { getTrailHeads } = require('./utilities/helper');
 const { fireWeatherApi } = require('./utilities/helper');
 const { getWeatherData } = require('./utilities/helper');
 const { updateWeatherStations } = require('./utilities/updateWeatherStations');
+const { getRainData } = require ('./utilities/rainData.js')
 
 //CONSTANTS
 const PORT = process.env.PORT  || 3000;
@@ -34,4 +34,6 @@ app.listen(PORT, () => {
   console.log(`SERVER IS LISTENING ON ${PORT}`);  
   // getTrailHeads();
   // updateWeatherStations();
+  // getRainData();
+
 });
